Got saving/loading the map from file to work
Had to refactor the Map and Node classes so that they were serialazable by the Json Utility. Also had to adjust things that used those object classes so that they worked with the new versions, and adjusted how sprites are loaded so that it would work with the saving/loading, because apparently sprites are assigned when unity runs or something, and may end up being different from the sprite references saved in the .json object.