Better corner drawing and node connection handling.
Did a whole bunch of stuff.
Updated nodes to handle connections and line data differently, required making a function that would load maps and nodes in the old version, and then convert them to the new version so they could be saved.
Also made corner drawing better, so now corners don't disappear when walking through a one-way connection.
now when a one-way connection would severe  the string, it does not draw it, and prevents you from crossing that connetion.
Rooms in the LevelsEdited folder have not been updated to the new map version. I can do that, or tell you how to do that if needed.